fix(db): atualiza modelo físico revisado pelo professor (corrige id_cliente em pedido, mantém enums/índices)

+++ banco/schema.sql
@@ -23,77 +23,70 @@ END $$;
 
 -- Cliente (1)───(N) Reserva; (1)───(N) Pedido
 CREATE TABLE IF NOT EXISTS cliente (
-  id_cliente   BIGSERIAL PRIMARY KEY,
-  nome         VARCHAR(120)  NOT NULL,
-  email        VARCHAR(254)  NOT NULL UNIQUE,
-  telefone     VARCHAR(20)
+  id_cliente SERIAL PRIMARY KEY,
+  nome       VARCHAR(120)  NOT NULL,
+  email      VARCHAR(254)  NOT NULL UNIQUE,
+  telefone   VARCHAR(20)
 );
 
 -- Mesa (1)───(N) Reserva
 CREATE TABLE IF NOT EXISTS mesa (
-  id_mesa     BIGSERIAL PRIMARY KEY,
-  numero      INTEGER     NOT NULL UNIQUE,
-  capacidade  SMALLINT    NOT NULL CHECK (capacidade > 0)
+  id_mesa     SERIAL PRIMARY KEY,
+  numero      INTEGER   NOT NULL UNIQUE,
+  capacidade  SMALLINT  NOT NULL CHECK (capacidade > 0)
 );
 
 -- CardapioItem (lado N da N:N com Pedido via Pedido_Item)
 CREATE TABLE IF NOT EXISTS cardapioitem (
-  id_item    BIGSERIAL PRIMARY KEY,
+  id_item    SERIAL PRIMARY KEY,
   nome       VARCHAR(120) NOT NULL,
   descricao  TEXT,
   preco      NUMERIC(10,2) NOT NULL CHECK (preco >= 0),
   categoria  VARCHAR(30)   NOT NULL,
   CONSTRAINT chk_cardapio_categoria
     CHECK (categoria IN ('bebida','sobremesa','lanche','outro'))
-  -- Se desejar, pode tornar o nome único:
-  -- , UNIQUE (nome)
 );
 
 -- Reserva (N)───(1) Mesa, (N)───(1) Cliente
 CREATE TABLE IF NOT EXISTS reserva (
-  id_reserva   BIGSERIAL PRIMARY KEY,
-  inicio       TIMESTAMP   NOT NULL,
-  fim          TIMESTAMP   NOT NULL,
-  qtd_pessoas  SMALLINT    NOT NULL CHECK (qtd_pessoas > 0),
+  id_reserva   SERIAL PRIMARY KEY,
+  inicio       TIMESTAMP  NOT NULL,
+  fim          TIMESTAMP  NOT NULL,
+  qtd_pessoas  SMALLINT   NOT NULL CHECK (qtd_pessoas > 0),
   status       reserva_status NOT NULL DEFAULT 'pendente',
 
   -- Decisão de negócio: ao excluir o cliente, apagar suas reservas
-  id_cliente   BIGINT NOT NULL REFERENCES cliente(id_cliente) ON DELETE CASCADE,
+  id_cliente   INTEGER NOT NULL REFERENCES cliente(id_cliente) ON DELETE CASCADE,
   -- Não permitir excluir mesa se houver reservas referenciando-a
-  id_mesa      BIGINT NOT NULL REFERENCES mesa(id_mesa)     ON DELETE RESTRICT,
+  id_mesa      INTEGER NOT NULL REFERENCES mesa(id_mesa)       ON DELETE RESTRICT,
 
   CHECK (fim > inicio)
-  -- Opcional: evita a mesma mesa no mesmo intervalo EXATO
+  -- Opcional: evitar mesma mesa no mesmo intervalo EXATO
   -- , UNIQUE (id_mesa, inicio, fim)
 );
 
 -- Pedido (N)───(1) Cliente
 CREATE TABLE IF NOT EXISTS pedido (
-  id_pedido   BIGSERIAL PRIMARY KEY,
+  id_pedido   SERIAL PRIMARY KEY,
   observacao  TEXT,
   status      pedido_status NOT NULL DEFAULT 'em_preparo',
   data_hora   TIMESTAMP     NOT NULL DEFAULT now(),
 
-  -- Decisão de negócio: ao excluir o cliente, apagar seus pedidos
-  id_cliente  BIGINT NOT NULL REFERENCES cliente(id_cliente) ON DELETE CASCADE
+  -- Corrigido: id_cliente (sem 'I' no final)
+  id_cliente  INTEGER NOT NULL REFERENCES cliente(id_cliente) ON DELETE CASCADE
 );
 
--- Pedido_Item (tabela de junção N:N entre Pedido e CardapioItem)
--- (padronização de nomes: usar id_item, igual à PK de cardapioitem)
+-- Pedido_Item (N:N Pedido x CardapioItem)
 CREATE TABLE IF NOT EXISTS pedido_item (
-  id_pedido  BIGINT NOT NULL REFERENCES pedido(id_pedido)         ON DELETE CASCADE,
-  id_item    BIGINT NOT NULL REFERENCES cardapioitem(id_item)     ON DELETE RESTRICT,
+  id_pedido  INTEGER NOT NULL REFERENCES pedido(id_pedido)         ON DELETE CASCADE,
+  id_item    INTEGER NOT NULL REFERENCES cardapioitem(id_item)     ON DELETE RESTRICT,
   quantidade INTEGER NOT NULL CHECK (quantidade > 0),
-
   PRIMARY KEY (id_pedido, id_item)
 );
 
--- 4) Índices auxiliares para FKs
+-- 4) Índices auxiliares (FKs)
 CREATE INDEX IF NOT EXISTS idx_reserva_id_cliente   ON reserva (id_cliente);
 CREATE INDEX IF NOT EXISTS idx_reserva_id_mesa      ON reserva (id_mesa);
 CREATE INDEX IF NOT EXISTS idx_pedido_id_cliente    ON pedido  (id_cliente);
 CREATE INDEX IF NOT EXISTS idx_pedidoitem_id_pedido ON pedido_item (id_pedido);
 CREATE INDEX IF NOT EXISTS idx_pedidoitem_id_item   ON pedido_item (id_item);
-
-
--- FIM
